+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+* Configure the GDS API Adapters logger to use logstasher
+* More consistent log level configuration by default
+
 # 1.12.0
 
 * Make ActiveRecord healthcheck more accurate